general refactor
(cherry picked from commit 1979df8d69914031957744e8dcaecf571719f561)
