+++ src/js/menu.js
@@ -2,8 +2,8 @@ const openAboutWindow = require("about-window").default;
 const path = require("path");
 
 const appInfo = require("../../package.json");
-const config = require("../config");
-const { CONSTANTS } = require("../util");
+const config = require("../helpers/config");
+const { CONSTANTS } = require("../helpers/util");
 
 const about = () => {
   openAboutWindow({
