+++ src/templates/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <title>Google Drive</title>
-    <link rel="stylesheet" href="../stylesheets/base.css" />
+    <link rel="stylesheet" href="../css/base.css" />
   </head>
 
   <body>
