--- src/util.js
@@ -1,19 +1,0 @@
-const config = require("./config");
-
-const CONSTANTS = {
-  OS_PLATFORMS: {
-    MAC_OS: "darwin",
-    WINDOWS: "win32",
-    LINUX: "linux",
-    AIX: "aix",
-    FREEBSD: "freebsd",
-    OPENBSD: "openbsd",
-    SUN_OS: "sunos",
-  },
-};
-
-// Use frameless title shift only on MacOS.Use os specific titlebar for other OS's.
-const TITLE_BAR_HEIGHT =
-  config.osPlatform === CONSTANTS.OS_PLATFORMS.MAC_OS ? 20 : 0;
-
-module.exports = { TITLE_BAR_HEIGHT, CONSTANTS };
